backend/requirements.txt
fastapi
langchain-community
langchain
uvicorn
openai
python-dotenv
langchain-openai
python-multipart